kernel: Boost DDR bus for a short period when zygote forks
There is a gap from when a new app is forked to the time that the
devfreq governor managing the DDR bus can pick up the increased load and
ramp up appropriately. Since the polling interval for this governor is
typically 50 ms, that means it may take up to 50 ms before the governor
ramps up, which hurts cold launch app performance. Boosting the DDR bus
for 50 ms upon a fork from zygote covers that gap, speeding up cold app
launches.

[ghostrider-reborn]
- Add input timeout of 2s since we don't want to boost DDR for
  zygote forks not initiated by user input, such as services
  spawning in background

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>