drm: Boost DDR bus upon running an atomic ioctl
In order to reduce jank, request a DDR bus boost whenever an atomic
ioctl is sent, which is what our display HAL uses to commit frames to
the display. The boost should be sufficient enough to render the
display's maximum FPS without any dropped frames when there is no
significant external source of load.

[ghostrider-reborn]:
- Boost if only within 3.2s input timeout
- This prevents unnecessary boosting such as during video playback
- More details: caf-ginkgo/kernel_msm-4.14@4ccf726

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>